refactor: clean conversation models and validations
