+++ conversation_service/models/__init__.py
@@ -1,56 +1,43 @@
-from .agent_models import AgentStep, AgentTrace
-from .conversation_models import Conversation, ConversationSummary, ConversationTurn
-
-__all__ = [
-    "AgentStep",
-    "AgentTrace",
-    "Conversation",
-    "ConversationSummary",
-    "ConversationTurn",
-from .conversation_models import (
-    ConversationRequest,
-    ConversationResponse,
-    ConversationMetadata,
-    ConversationContext,
-)
-from .conversation_db_models import (
 """Expose Pydantic models for external use."""
 
 from .agent_models import (
     AgentConfig,
-    IntentResult,
-    DynamicFinancialEntity,
     AgentResponse,
+    AgentStep,
+    AgentTrace,
+    DynamicFinancialEntity,
+    IntentResult,
 )
-from .conversation_models import (
+from .conversation_db_models import (
     Conversation,
     ConversationSummary,
     ConversationTurn,
 )
-
-__all__ = [
-    "ConversationRequest",
-    "ConversationResponse",
-    "ConversationMetadata",
-    "ConversationContext",
-    "Conversation",
-    "ConversationSummary",
-    "ConversationTurn",
-    "AgentConfig",
-    "IntentResult",
-    "DynamicFinancialEntity",
-    "AgentResponse",
-    "Conversation",
-    "ConversationSummary",
-    "ConversationTurn",
-
-"""Exports publics du package `conversation_service.models`."""
-
+from .conversation_models import (
+    ConversationContext,
+    ConversationMetadata,
+    ConversationRequest,
+    ConversationResponse,
+)
 from .enums import ConfidenceThreshold, EntityType, IntentType, QueryType
 
 __all__ = [
-    "IntentType",
-    "EntityType",
-    "QueryType",
+    "AgentConfig",
+    "AgentResponse",
+    "AgentStep",
+    "AgentTrace",
+    "DynamicFinancialEntity",
+    "IntentResult",
+    "Conversation",
+    "ConversationSummary",
+    "ConversationTurn",
+    "ConversationContext",
+    "ConversationMetadata",
+    "ConversationRequest",
+    "ConversationResponse",
     "ConfidenceThreshold",
+    "EntityType",
+    "IntentType",
+    "QueryType",
 ]
+
